+++ src/entities/cloud.py
