+++ src/entities/coin.py
