+++ src/entities/vegetable.py
