+++ src/maps_engine.py
@@ -3,15 +3,15 @@ import os
 
 from pygame import Rect,sprite,mixer
 
-from src.entitys.vegetable import Vegetable
+from src.entities.vegetable import Vegetable
 from src.blocks.lootBlock import LootBlock
 from src.blocks.platform import Platform
-from src.entitys.goomba import Goomba
-from src.entitys.player import Player
-from src.entitys.koopa import Koopa
-from src.entitys.cloud import Cloud
+from src.entities.goomba import Goomba
+from src.entities.player import Player
+from src.entities.koopa import Koopa
+from src.entities.cloud import Cloud
 from src.blocks.floor import Floor
-from src.entitys.coin import Coin
+from src.entities.coin import Coin
 
 
 class Camera(object):
